workspaces/trading-bot/requirements.txt
ccxt>=4.2.0
PyYAML>=6.0.1
pandas>=2.1.0
numpy>=1.26.0
redis>=5.0.0
aiosqlite>=0.20.0
aiohttp>=3.9.0
